+++ vulnerabilities/models.py
@@ -218,6 +218,19 @@ class PackageQuerySet(BaseQuerySet, PackageURLQuerySet):
         )
 
 
+def get_purl_query_lookups(purl):
+    """
+    Do not reference all the possible qualifiers and relax the
+    purl matching to only lookup the type, namespace, name and version fields.
+    """
+    lookup_fields = ["type", "namespace", "name", "version"]
+    return {
+        field_name: value
+        for field_name, value in purl.to_dict().items()
+        if value and field_name in lookup_fields
+    }
+
+
 class Package(PackageURLMixin):
     """
     A software package with related vulnerabilities.